requirements.txt
Babel==2.1.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
Flask-WhooshAlchemy==0.56
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Tempita==0.5.2
WTForms==2.0.2
Werkzeug==0.10.4
Whoosh==2.7.0
blinker==1.4
coverage==4.0.3
decorator==4.0.4
defusedxml==0.4.1
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
pbr==1.8.0
psycopg2==2.6.1
python3-openid==3.0.7
pytz==2015.6
rauth==0.7.1
requests==2.8.0
six==1.9.0
speaklater==1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
